+++ example/lib/game_config.dart
@@ -1,7 +1,9 @@
 import 'package:bishop/bishop.dart' as bishop;
+import 'package:squares/squares.dart';
 
 class GameConfig {
   final bishop.Variant variant;
+  final int humanPlayer;
 
-  GameConfig({required this.variant});
+  GameConfig({required this.variant, this.humanPlayer = WHITE});
 }